Move duck-typing range filter into case search DSL
Everything else should already know whether it's a numeric or date range
