+++ corehq/apps/es/case_search.py
@@ -11,11 +11,7 @@ from corehq.apps.es import case_search as case_search_es
 """
 
 from copy import deepcopy
-from datetime import date, datetime
-from warnings import warn
-
-from django.utils.dateparse import parse_date, parse_datetime
-from django.utils.translation import gettext
+from datetime import datetime
 
 from memoized import memoized
 
@@ -267,27 +263,6 @@ def case_property_starts_with(case_property_name, value):
     )
 
 
-def case_property_range_query(case_property_name, gt=None, gte=None, lt=None, lte=None):
-    """Returns cases where case property `key` fall into the range provided."""
-    kwargs = {'gt': gt, 'gte': gte, 'lt': lt, 'lte': lte}
-    try:
-        # if its a number, use it
-        kwargs = {key: float(value) for key, value in kwargs.items() if value is not None}
-    except (TypeError, ValueError):
-        pass
-    else:
-        return case_property_numeric_range(case_property_name, **kwargs)
-
-    kwargs = {
-        key: value if isinstance(value, (date, datetime)) else _parse_date_or_datetime(value)
-        for key, value in kwargs.items()
-        if value is not None
-    }
-    if not kwargs:
-        raise TypeError()       # Neither a date nor number was passed in
-    return case_property_date_range(case_property_name, **kwargs)
-
-
 def case_property_numeric_range(case_property_name, gt=None, gte=None, lt=None, lte=None):
     kwargs = {'gt': gt, 'gte': gte, 'lt': lt, 'lte': lte}
     return _base_property_query(
@@ -302,30 +277,6 @@ def case_property_date_range(case_property_name, gt=None, gte=None, lt=None, lte
         case_property_name,
         queries.date_range("{}.{}.date".format(CASE_PROPERTIES_PATH, VALUE), **kwargs)
     )
-
-
-def _parse_date_or_datetime(value):
-    parsed_date = _parse_date(value)
-    if parsed_date is not None:
-        return parsed_date
-    parsed_datetime = _parse_datetime(value)
-    if parsed_datetime is not None:
-        return parsed_datetime
-    raise ValueError(gettext(f"{value} is not a correctly formatted date or datetime."))
-
-
-def _parse_date(value):
-    try:
-        return parse_date(value)
-    except ValueError:
-        raise ValueError(gettext(f"{value} is an invalid date."))
-
-
-def _parse_datetime(value):
-    try:
-        return parse_datetime(value)
-    except ValueError:
-        raise ValueError(gettext(f"{value} is an invalid datetime."))
 
 
 def reverse_index_case_query(case_ids, identifier=None):
